lab4/server/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.0
environ==1.0
idna==3.3
Pillow==8.4.0
pycodestyle==2.8.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
psycopg2>=2.8
djangorestframework==3.13.1
